% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_dados.R
\name{baixar_trimestres_pnadc}
\alias{baixar_trimestres_pnadc}
\title{Baixar e consolidar os 4 trimestres de um determinado ano}
\usage{
baixar_trimestres_pnadc(
  ano,
  vars_tri = vars_tri_default,
  low_memory = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ano}{Ano de referencia (inteiro).}

\item{vars_tri}{Vetor de nomes de variaveis a serem baixadas.}

\item{low_memory}{Logico. Se TRUE, salva intermediarios em disco temporario.}

\item{verbose}{Logico. Se TRUE, exibe mensagens de progresso.}
}
\value{
Data frame consolidado das pessoas nos 4 trimestres.
}
\description{
Baixar e consolidar os 4 trimestres de um determinado ano
}
\keyword{internal}
