% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downcast.R
\name{downcast_pnadc}
\alias{downcast_pnadc}
\title{Downcast PNADc data types}
\usage{
downcast_pnadc(df)
}
\arguments{
\item{df}{Data frame / tibble contendo microdados da PNADc.}
}
\value{
Data frame com colunas categoricas/inteiras convertidas para tipo integer.
}
\description{
Converte colunas numericas/categoricas dos microdados da PNADc para inteiros de 32 bits
para otimizar o uso da memoria RAM durante o processamento.
}
\keyword{internal}
