% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_wrapper.R
\name{get_pnadc_internal}
\alias{get_pnadc_internal}
\title{Wrapper Interno de Download da PNADc com Suporte a Mocks e Retry}
\usage{
get_pnadc_internal(
  year,
  quarter = NULL,
  interview = NULL,
  vars = NULL,
  design = FALSE,
  labels = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Ano de referencia.}

\item{quarter}{Trimestre (1-4) para dados trimestrais, ou NULL.}

\item{interview}{Entrevista (1-5) para dados de Visita, ou NULL.}

\item{vars}{Vetor de variaveis a baixar.}

\item{design}{Logico.}

\item{labels}{Logico.}

\item{verbose}{Logico.}
}
\value{
Data frame com os microdados baixados.
}
\description{
Wrapper Interno de Download da PNADc com Suporte a Mocks e Retry
}
\keyword{internal}
