% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_dados.R
\name{vars_tri_default}
\alias{vars_tri_default}
\title{Variaveis padrao trimestrais da PNAD Contínua}
\usage{
vars_tri_default
}
\description{
Lista contendo as variaveis padrao da base trimestral utilizadas para
construcao do painel de pessoas.
}
