% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_dados.R
\name{vars_visita_default}
\alias{vars_visita_default}
\title{Variaveis padrao de Visita 1 (Domicilio) da PNAD Contínua}
\usage{
vars_visita_default
}
\description{
Lista contendo as variaveis padrao da base de Visita 1 utilizadas para
caracterizacao domiciliar e rendimento per capita.
}
